Implemented function _get_root_domain_name.
Currently, the authenticate function only allows creation of certificates for root domain name because the authenticate function compares the hostname with the root domain name. Comparing the root domain name of the hostname with the root domain name can resolve this issue. Implemented the _get_root_domain_name function which is now used in the authenticate function.